Move refreshable patch to installer
