+++ source/SongCore/Installers/AppInstaller.cs
@@ -1,14 +1,51 @@
-using SongCore.HarmonyPatches;
+using System;
+using System.Collections.Concurrent;
 using Zenject;
 
 namespace SongCore.Installers
 {
     internal class AppInstaller : Installer
     {
+        private const string refreshableID = "SongCore.Loader.Refresh";
+        private const string didLoadEventID = "SongCore.Loader.Loaded";
+
         public override void InstallBindings()
         {
-            // TODO
-            MainSystemsInitRefreshablePatch.Postfix(Container);
+            Container.Bind<IRefreshable>().WithId(refreshableID).To<SongCoreRefreshable>().AsSingle();
+            Container.Bind(typeof(IInitializable), typeof(IDisposable), typeof(SongCoreLoaderDidLoad)).To<SongCoreLoaderDidLoad>().AsSingle();
+            var loadEvent = Container.Resolve<SongCoreLoaderDidLoad>() as IObservableChange;
+            Container.BindInstance(loadEvent).WithId(didLoadEventID).AsSingle();
+        }
+
+        private class SongCoreRefreshable : IRefreshable
+        {
+            public void Refresh()
+            {
+                if (Loader.AreSongsLoaded)
+                {
+                    Loader.Instance.RefreshSongs();
+                }
+            }
+        }
+
+        private class SongCoreLoaderDidLoad : IInitializable, IDisposable, IObservableChange
+        {
+            public event Action? didChangeEvent;
+
+            public void Initialize()
+            {
+                Loader.SongsLoadedEvent += Loader_SongsLoadedEvent;
+            }
+
+            private void Loader_SongsLoadedEvent(Loader _, ConcurrentDictionary<string, BeatmapLevel> __)
+            {
+                didChangeEvent?.Invoke();
+            }
+
+            public void Dispose()
+            {
+                Loader.SongsLoadedEvent -= Loader_SongsLoadedEvent;
+            }
         }
     }
 }
